Convert postModel to ts and make explore page default route
To be able to reference a TypeScript-defined interface for pulling in
post objects from the API, I converted postModel.js to use ts instead.
This involved some changes to how to run the backend, as updated in
`run.sh`. Everything still works as before, but the type safety added by
TS will make the rest of the work on the explore page easier.